chore: various updates (#16)
This pr just updates a handful of stuff. The reason is that it was
failing in the [Steward
Runs](https://github.com/VirtusLab/scala-steward-repos/issues/98#issuecomment-1366745022).
So this should  ensure that the steward plugin doesn't have any issues.

Note that I do rip out some of the custom stuff you have for sonatype
publishing and just replace it with
https://github.com/ckipp01/mill-ci-release mainly because it does almost
exactly what you were doing in here. Let me know if that's not ok and I
can revert it.